signed/unsigned comparison warning fix
git-svn-id: https://ricamsvn.ricam.oeaw.ac.at/gismo/stable@5134 e128ff1e-fa79-4151-a8cb-522299a2c450

+++ src/gsCore/gsMultiBasis.h
@@ -288,7 +288,7 @@ public:
     /// on patch \em k with extension \em refExt.
     ///
     /// See gsHTensorBasis::refineWithExtension() for further documentation.
-    void refine(int k, gsMatrix<T> const & boxes, int refExt)
+    void refine(size_t k, gsMatrix<T> const & boxes, int refExt)
     {
         GISMO_ASSERT( k < m_bases.size(),
                       "Invalid patch index "<<k<<" requested from gsMultiBasis" );
